requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
reportlab
itsdangerous
Flask-Mail
Flask-WTF
reportlab
python-dotenv
gunicorn
